Stunt sytem preperations, car resetting,  jack car, car selection menu
Made preparations to include the stunting system, found out that the car needs to be updated in order for it to work, removed unnecessary garbage bins under the map, tested stunt system using the implemented car from Jack which worked, cleaned up some stuff using new made folders, added car resetting in the car controller script and changed the keyboard input to controller input (needs to be checked) and tested this which worked. Changed an unused scene to a simple car selection menu (for now) which works (except for selecting a car as no cars are available).